requirements.txt
fastapi==0.104.1
uvicorn==0.23.2

requests==2.31.0
beautifulsoup4==4.12.2

langchain==0.0.335
langchain-community==0.0.13

pydantic==2.4.2
python-dotenv==1.0.0

youtube-transcript-api==1.0.3
google-generativeai>=0.3.0